remove plugins. replace mocha with tape. modify README. modify CHANGELOG. use version 0.0.1. add debuglog to index.js
