+++ lib/middleware.js
@@ -24,7 +24,7 @@ var fs = require('fs'),
     mkdirp = require('mkdirp'),
     merge = require('n-deep-merge'),
 	noop = require('./noop'),
-	debuglog = require('debuglog')('kraken-devtools'),
+	debuglog = require('debuglog')('construx'),
 	filter = require('./filter');
