--- lib/requireany.js
@@ -1,42 +1,0 @@
-/*───────────────────────────────────────────────────────────────────────────*\
- │  Copyright (C) 2014 eBay Software Foundation                                │
- │                                                                             │
- │hh ,'""`.                                                                    │
- │  / _  _ \  Licensed under the Apache License, Version 2.0 (the "License");  │
- │  |(@)(@)|  you may not use this file except in compliance with the License. │
- │  )  __  (  You may obtain a copy of the License at                          │
- │ /,'))((`.\                                                                  │
- │(( ((  )) ))    http://www.apache.org/licenses/LICENSE-2.0                   │
- │ `\ `)(' /'                                                                  │
- │                                                                             │
- │   Unless required by applicable law or agreed to in writing, software       │
- │   distributed under the License is distributed on an "AS IS" BASIS,         │
- │   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  │
- │   See the License for the specific language governing permissions and       │
- │   limitations under the License.                                            │
- \*───────────────────────────────────────────────────────────────────────────*/
-'use strict';
-
-
-var assert = require('assert');
-
-
-module.exports = function requireAny(/*modules*/) {
-    var result, failed;
-
-    result = undefined;
-    failed = [];
-
-    Array.prototype.slice.call(arguments).some(function (moduleName) {
-        try {
-            result = require(moduleName);
-        } catch (err) {
-            // noop
-            failed.push(moduleName);
-        }
-        return !!result;
-    });
-
-    assert.ok(failed.length !== arguments.length, 'Required module(s) not found. Please install one of the following: ' + failed.join(', '));
-    return result;
-};
