--- plugins/sass.js
@@ -1,42 +1,0 @@
-/*───────────────────────────────────────────────────────────────────────────*\
- │  Copyright (C) 2014 eBay Software Foundation                                │
- │                                                                             │
- │hh ,'""`.                                                                    │
- │  / _  _ \  Licensed under the Apache License, Version 2.0 (the "License");  │
- │  |(@)(@)|  you may not use this file except in compliance with the License. │
- │  )  __  (  You may obtain a copy of the License at                          │
- │ /,'))((`.\                                                                  │
- │(( ((  )) ))    http://www.apache.org/licenses/LICENSE-2.0                   │
- │ `\ `)(' /'                                                                  │
- │                                                                             │
- │   Unless required by applicable law or agreed to in writing, software       │
- │   distributed under the License is distributed on an "AS IS" BASIS,         │
- │   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  │
- │   See the License for the specific language governing permissions and       │
- │   limitations under the License.                                            │
- \*───────────────────────────────────────────────────────────────────────────*/
-'use strict';
-
-
-var lib = require('node-sass');
-
-
-module.exports = function (options) {
-
-    options.ext = options.ext || 'scss';
-
-    return function (data, args, callback) {
-        lib.render({
-            data: data.toString(),
-            success: function success(result) {
-                callback(null, result.css || result); // result.css for ^2
-            },
-            error: function error(err) {
-                err.status = 500; // for ^2
-                callback(err);
-            },
-            includePaths: args.paths
-        });
-    };
-
-};
