--- test/fixtures/include/sass/elsewhere.scss
@@ -1,3 +1,0 @@
-body {
-    color: black;
-}
