--- test/fixtures/locales/US/es/localized.properties
@@ -1,1 +1,0 @@
-localized.greeting =  Hola {name}
