--- test/fixtures/public/css/less/app.less
@@ -1,8 +1,0 @@
-@import "inc/colors";
-
-#header {
-  color: @color;
-}
-h2 {
-  color: @color;
-}
