--- test/fixtures/public/css/less/inc/colors.less
@@ -1,5 +1,0 @@
-@color: #4D926F;
-
-a:hover {
-  color: @color;
-}
