--- test/fixtures/public/css/less/invalid.less
@@ -1,5 +1,0 @@
-@color #4D926F;
-
-a:hover {
-  color: @color;
-}
