--- test/fixtures/public/css/less/missing.less
@@ -1,1 +1,0 @@
-@import "inc/notfound";
