--- test/fixtures/public/css/sass/app.scss
@@ -1,16 +1,0 @@
-@import "inc/colors";
-
-$blue: #3bbfce;
-$margin: 16px;
-
-.content-navigation {
-  border-color: $blue;
-  color:
-    darken($blue, 9%);
-}
-
-.border {
-  padding: $margin / 2;
-  margin: $margin / 2;
-  border-color: $color;
-}
