--- test/fixtures/public/css/sass/inc/colors.scss
@@ -1,5 +1,0 @@
-$color: #4D926F;
-
-a:hover {
-  color: $color;
-}
