--- test/fixtures/public/css/sass/includePath.scss
@@ -1,1 +1,0 @@
-@import "elsewhere";
