--- test/fixtures/public/css/sass/invalid.scss
@@ -1,5 +1,0 @@
-$color #4D926F;
-
-a:hover {
-  color: $color;
-}
