--- test/fixtures/public/css/sass/missing.scss
@@ -1,1 +1,0 @@
-@import "inc/notfound";
