--- test/fixtures/public/css/stylus/app.styl
@@ -1,13 +1,0 @@
-border-radius()
-  -webkit-border-radius: arguments
-  -moz-border-radius: arguments
-  border-radius: arguments
-
-body a
-  font: 12px/1.4 "Lucida Grande", Arial, sans-serif
-  background: black
-  color: #ccc
-
-form input
-  padding: 5px
-  border: 1px solid
