--- test/fixtures/public/css/stylus/missing.styl
@@ -1,1 +1,0 @@
-@import 'inc/notfound'
