--- test/fixtures/public/img/altfile
@@ -1,1 +1,0 @@
-File intentionally left blank.
