--- test/fixtures/public/img/omg+path/omg+file.png
@@ -1,1 +1,0 @@
-pluspng
