--- test/fixtures/public/img/omg.path/index.html
@@ -1,1 +1,0 @@
-cloud easy
