--- test/fixtures/public/img/omgpath/index.html
@@ -1,1 +1,0 @@
-cloud easy
