--- test/fixtures/public/js/main.js
@@ -1,4 +1,0 @@
-(function () {
-    'use strict';
-
-}());
