--- test/fixtures/public/templates/inc/partial.dust
@@ -1,1 +1,0 @@
-<h1>Hello, world!</h1>
