--- test/fixtures/public/templates/index.dust
@@ -1,1 +1,0 @@
-<h1>Hello, world!</h1>
