--- test/fixtures/public/templates/invalid.dust
@@ -1,1 +1,0 @@
-<h1>Hello, {@name/ }!</h1>
