--- test/fixtures/public/templates/localized.dust
@@ -1,1 +1,0 @@
-<div>{@pre type="content" key="localized.greeting"/}</div>
